+++ Config/app.js
@@ -1,3 +1,4 @@
+// Biblioteca principal
 const http = require('http')
 
 const { createTable, readColumn } = require('./database/db')
@@ -49,6 +50,7 @@ const server = http.createServer( async (req, res) => {
             // -> Esses dados são capturados pela fetch API (visto no script Cliente.JS) e renderizados na DOM do projeto (home.js, meu-perfil.js...) .
 
         let id = fetchID() 
+        let query 
 
         if (req.url == '/api/user') {
             query = `SELECT * FROM dados_clientes where cpf = '${id}'`;
@@ -62,8 +64,8 @@ const server = http.createServer( async (req, res) => {
             query = `SELECT * FROM conta WHERE cpf = '${id}'`;
         }
         const database = await readColumn(query)
-        console.log('PUXANDO QUERY... DO DB', database.rows, req.url)
-        const userData = database.rows 
+        const userData = database.rows
+
         res.writeHead(200, { 'Content-Type': 'application/json', 'Access-Control-Allow-Origin': '*'})
         res.end(JSON.stringify(userData))
