+++ Config/scripts/getResponse.js
@@ -1,26 +1,16 @@
 
-const { createColumn, readColumn, updateColumn } = require('../database/db')
-const { sendEmail } = require('../scripts/sendEmail')
-const { transferCash, checkCPF } = require('./transferOp')
-const { getError } = require('./getError')
+const { readColumn, updateColumn } = require('../database/db')
+const { getError } = require('../database/getError')
+
+const { sendEmail } = require('../modules/sendEmail')
+const { transferCash } = require('../modules/transferOp')
 
 // Funções que interagem com banco e recebem as requisições do "Front-End" e voltam uma resposta para a página
 function getResponse() {
 
-    // Função para retornar a resposta com OUTCOME (200/400) e resposta customizada
-    function returnResponse(res, outcome, response) {
-        res.writeHead(outcome, { 'Content-Type': 'application/json', 'Access-Control-Allow-Origin': '*' })
-        res.end(JSON.stringify({
-            message: response
-        }))
-    }
-
-    /* ============================================================================================================== */
-
-
     // Função para salvar cadastro de usuário
     const saveData = async (data, res) => {
-        const operation = await createColumn(`INSERT INTO dados_clientes (nome, cpf, email, telefone, data_nascimento, senha) VALUES ($1, $2, $3, $4, $5, $6)`, [data.nome, data.cpf, data.email, data.telefone, data.data_nascimento, data.senha]).catch((err) => { console.log(err) })
+        const operation = await updateColumn(`INSERT INTO dados_clientes (nome_completo, cpf, email, telefone, data_de_nascimento, senha) VALUES ($1, $2, $3, $4, $5, $6)`, [data.nome, data.cpf, data.email, data.telefone, data.data_nascimento, data.senha]).catch((err) => { console.log(err) })
         const response = operation.outcome == 400 ? getError(operation.error) : 'Cadastro concluído com sucesso!'
         returnResponse(res, operation.outcome, response)
     }
@@ -58,31 +48,44 @@ function getResponse() {
             
             returnResponse(res, outcome, "")
         
-
         }
     }
 
     // Função para lidar com dados "avulsos", não referentes ao cadastro
-    const sendData = (data, res, fetchID) => {
+    const sendData =  async (data, res, fetchID) => {
         
         let operation 
         
         if (data.type == 'send_email') {
-            operation = sendEmail(data.email)
+            operation = await sendEmail(data.email)
         }
         
         else if (data.type == 'transfer_cash') {
-            operation = transferCash(data, fetchID)
+            operation = await transferCash(data, fetchID)
         }
 
+        
+        returnResponse(res, operation.outcome, operation.response)
+        /*
         operation.then((result) => {
             res.writeHead(result.outcome, { 'Content-Type': 'application/json', 'Access-Control-Allow-Origin': '*' })
             res.end(JSON.stringify({
                 message: result.response
             }))
         })
+        */
 
     }   
+
+
+    // Função para retornar a resposta com OUTCOME (200/400) e resposta customizada
+    function returnResponse(res, outcome, response) {
+        res.writeHead(outcome, { 'Content-Type': 'application/json', 'Access-Control-Allow-Origin': '*' })
+        res.end(JSON.stringify({
+            message: response
+        }))
+    }
+
 
     return { saveData, validateData, updateData, sendData }
 }
