+++ Models/extrato.html
@@ -43,7 +43,7 @@
     <nav id="menu">
 
         <header >
-            <a href="#" class="btn-expandir" onclick="abrirMenu()" ><img src="./imgs/capi.png" alt=""></a>
+            <a href="#" class="btn-expandir" ><img src="./imgs/capi.png" alt=""></a>
         </header>
 
         <div class="part1">
@@ -85,7 +85,7 @@
                 <span class="txt-link">Ajuda</span>
             </a>
 
-            <a href="#" onclick="fecharMenu()">
+            <a href="#">
                 <span class="icon"><i class='bx bx-log-out-circle'></i></span>
                 <span class="txt-link">Diminuir</span>
             </a>
