+++ Models/pix.html
@@ -37,14 +37,14 @@
 					</div>
 					<div>
 						<label for="dispo">Valor Disponível:</label>
-						<input class="fundoCinza" type="text" name="dispo" id="dispo">
+						<input class="fundoCinza" type="text" name="dispo" id="dispo" disabled>
 					</div>
 				</div>
 				<div>
 					<label for="descri">Descrição</label>
 					<input type="text" name="descri" id="descri">
 				</div>
-				<div><label class="texto"> Instituição CAPYBANCK MS</label></div>
+				<div><label class="texto"> Instituição CAPYBANK MS</label></div>
 				<div class="btn">
 					<button type="submit" class="btn-enviar" id = "submit">ENVIAR</button>
 					<a href="home.html">CANCELAR</a>
